requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
MDAnalysis>=2.4.0
mdtraj>=1.9.7
PyYAML>=6.0
jinja2>=3.0.0
plotly>=5.0.0
kaleido>=0.2.1
scikit-learn>=1.0.0
networkx>=2.6.0
biopython>=1.79
pytest>=6.0.0
pytest-cov>=3.0.0 